fix a bug cause server crash.
