+++ src/main/java/com/github/alexthe666/iceandfire/compat/bauble/client/BaublesCompatBridgeClient.java
@@ -1,23 +1,13 @@
 package com.github.alexthe666.iceandfire.compat.bauble.client;
 
-import com.github.alexthe666.iceandfire.compat.bauble.common.PlayerWearBaubleEvent;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.entity.RenderPlayer;
-import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.item.Item;
 import net.minecraftforge.fml.common.Loader;
 
 import java.util.Map;
 
-public class BaublesCompatBridge {
+public class BaublesCompatBridgeClient {
     private static final String COMPAT_MOD_ID = "baubles";
-
-    public static boolean isWearSpecificBauble(EntityLivingBase player, Item bauble) {
-        if (Loader.isModLoaded(COMPAT_MOD_ID)) {
-            return PlayerWearBaubleEvent.arePlayerWearBaubles(player, bauble);
-        } else
-            return false;
-    }
 
     public static void loadBaublesClientModels() {
         if (Loader.isModLoaded(COMPAT_MOD_ID)) {
